// server/src/models/DrumCorps.ts
export enum DrumCorps {
  theAcademy = 'theAcademy',
  blueDevils = 'blueDevils',
  blueKnights = 'blueKnights',
  blueStars = 'blueStars',
  bluecoats = 'bluecoats',
  bostonCrusaders = 'bostonCrusaders',
  theCadets = 'theCadets',
  carolinaCrown = 'carolinaCrown',
  theCavaliers = 'theCavaliers',
  colts = 'colts',
  crossmen = 'crossmen',
  genesis = 'genesis',
  jerseySurf = 'jerseySurf',
  madisonScouts = 'madisonScouts',
  mandarins = 'mandarins',
  musicCity = 'musicCity',
  pacificCrest = 'pacificCrest',
  phantomRegiment = 'phantomRegiment',
  santaClaraVanguard = 'santaClaraVanguard',
  seattleCascades = 'seattleCascades',
  spiritOfAtlanta = 'spiritOfAtlanta',
  troopers = 'troopers',
}